requirements.txt
Flask 
pytesseract
pillow
pdf2image
pyspellchecker